backend/requirements.txt
Flask
PyMuPDF
python-dotenv
numpy
pandas
requests
pathway
litellm
transformers
google-generativeai
joblib
